Convert catfishing ID type to i32
cf_games.id, cf_articles.id and cf_articles.game_id are INTEGER (int4) in
Postgres. Switch CatfishingGame.id, CatfishingArticle.id, the CatfishingStore
id parameters and the api ListCatfishingGame DTO to i32, dropping the
CAST(... AS BIGINT) casts.

+++ api/src/catfishing.rs
@@ -21,7 +21,7 @@ impl From<CatfishingArticle> for ListCatfishingArticle {
 
 #[derive(Debug, Clone, Serialize)]
 pub struct ListCatfishingGame {
-    id: i64,
+    id: i32,
     articles: Vec<ListCatfishingArticle>,
 }
